chore: fixed cargo clippy warnings -windows

+++ src/utils/formatting.rs
@@ -24,14 +24,14 @@ pub fn format_bytes(size: u64) -> String {
 pub fn format_bytes_i64(size: i64) -> String {
     let abs_size = size.abs() as f64;
 
-    let formatted = if abs_size >= TB as f64 {
-        format!("{:.2} TB", abs_size / TB as f64)
-    } else if abs_size >= GB as f64 {
-        format!("{:.2} GB", abs_size / GB as f64)
-    } else if abs_size >= MB as f64 {
-        format!("{:.2} MB", abs_size / MB as f64)
-    } else if abs_size >= KB as f64 {
-        format!("{:.2} KB", abs_size / KB as f64)
+    let formatted = if abs_size >= TB {
+        format!("{:.2} TB", abs_size / TB)
+    } else if abs_size >= GB {
+        format!("{:.2} GB", abs_size / GB)
+    } else if abs_size >= MB {
+        format!("{:.2} MB", abs_size / MB)
+    } else if abs_size >= KB {
+        format!("{:.2} KB", abs_size / KB)
     } else {
         format!("{} B", abs_size as u64)
     };
